+++ src/main/java/nexusvault/cli/plugin/archive/NavigatorListDirectoryContentCmd.java
@@ -19,29 +19,6 @@ final class NavigatorListDirectoryContentCmd implements Command {
 		//@formatter:on
 	}
 
-	// @Override
-	// public void onCmd(CmdArgContainer args) {
-	// final ArchiveReader archive = resources.getArchive();
-	// if (archive == null) {
-	// throw new ArchiveNotSetException("Archive not set. Use 'help' to determine how to set the wildstar archive.");
-	// }
-	//
-	// final ArchivePath archivePath = resources.getInnerArchivePath();
-	// final IdxEntry selectedEntry = archivePath.resolve(archive.getRootFolder());
-	//
-	// if (selectedEntry.isFile()) {
-	// sendMsg("File: " + selectedEntry.fullName());
-	// } else {
-	// for (final IdxEntry child : selectedEntry.asDirectory().getChilds()) {
-	// if (child.isDir()) {
-	// sendMsg("Dir: " + child.fullName());
-	// } else if (child.isFile()) {
-	// sendMsg("File: " + child.fullName());
-	// }
-	// }
-	// }
-	// }
-
 	@Override
 	public void onCommand(CommandArguments args) {
 		App.getInstance().getPlugIn(ArchivePlugIn.class).listDirectoryContent();