+++ src/main/java/nexusvault/cli/plugin/export/ModelExporter.java
@@ -23,10 +23,9 @@ import java.util.Set;
 import javax.imageio.ImageIO;
 
 import nexusvault.archive.IdxDirectory;
-import nexusvault.archive.IdxEntryNotAFile;
-import nexusvault.archive.IdxEntryNotFound;
+import nexusvault.archive.IdxEntry;
 import nexusvault.archive.IdxFileLink;
-import nexusvault.archive.VaultReader;
+import nexusvault.archive.IdxPath;
 import nexusvault.archive.util.DataHeader;
 import nexusvault.cli.App;
 import nexusvault.cli.Command;
@@ -35,7 +34,7 @@ import nexusvault.cli.CommandInfo;
 import nexusvault.cli.ConsoleSystem.Level;
 import nexusvault.cli.model.ModelPropertyChangedEvent;
 import nexusvault.cli.plugin.archive.ArchivePlugIn;
-import nexusvault.cli.plugin.archive.SourcedVaultReader;
+import nexusvault.cli.plugin.archive.NexusArchiveWrapper;
 import nexusvault.format.m3.Model;
 import nexusvault.format.m3.export.gltf.GlTFExportMonitor;
 import nexusvault.format.m3.export.gltf.GlTFExporter;
@@ -111,7 +110,7 @@ final class ModelExporter implements Exporter {
 	}
 
 	public static interface InternalModelExporter {
-		void export(Model model, Path dstFolder, IdxFileLink fileLink) throws IOException;
+		void export(Model model, Path dstFolder, IdxPath filePath) throws IOException;
 	}
 
 	private ModelReader modelReader;
@@ -166,63 +165,46 @@ final class ModelExporter implements Exporter {
 	}
 
 	@Override
-	public void export(IdxFileLink fileLink, ByteBuffer data) throws IOException {
+	public void export(Path outputFolder, ByteBuffer data, IdxPath dataName) throws IOException {
 		final Model model = modelReader.read(data);
 
-		final Path outputFolder = App.getInstance().getAppConfig().getOutputPath();
-		final Path modelFolder = outputFolder.resolve(fileLink.fullName()).getParent();
+		final Path modelFolder = outputFolder.resolve(PathUtil.getFolder(dataName));
 		Files.createDirectories(modelFolder);
-		getExporter().export(model, modelFolder, fileLink);
+		getExporter().export(model, modelFolder, dataName);
 	}
 
 	private static final class GlTFInternalModelExporter implements InternalModelExporter {
-		private IdxFileLink findFile(IdxDirectory root, String fileName) {
-			try {
-				return root.getFile(fileName);
-			} catch (IdxEntryNotFound | IdxEntryNotAFile e) {
-				return null;
-			}
-		}
 
-		private SearchResult find(List<SourcedVaultReader> vaults, String textureId) {
-			for (final SourcedVaultReader vault : vaults) {
-				final IdxFileLink file = findFile(vault.getReader().getRootFolder(), textureId);
-				if (file != null) {
-					final SearchResult r = new SearchResult();
-					r.archive = vault.getReader();
-					r.fileLink = file;
-					return r;
+		private IdxFileLink find(List<NexusArchiveWrapper> wrappers, String textureId) {
+			for (final NexusArchiveWrapper wrapper : wrappers) {
+				final IdxDirectory root = wrapper.getArchive().getRootDirectory();
+				final IdxPath path = IdxPath.createPathFrom(textureId);
+				if (path.isResolvable(root)) {
+					final IdxEntry entry = path.resolve(root);
+					return entry.isFile() ? entry.asFile() : null;
 				}
 			}
 			return null;
 		}
 
-		private static class SearchResult {
-			public VaultReader archive;
-			public IdxFileLink fileLink;
-		}
-
 		@Override
-		public void export(Model model, Path dstFolder, IdxFileLink fileLink) throws IOException {
+		public void export(Model model, Path dstFolder, IdxPath filePath) throws IOException {
 			final GlTFExporter gltfExporter = new nexusvault.format.m3.export.gltf.GlTFExporter();
-			final String modelName = fileLink.getName().substring(0, fileLink.getName().lastIndexOf('.'));
+			final String modelName = PathUtil.getNameWithoutExtension(filePath);
 
-			final List<SourcedVaultReader> vaults = App.getInstance().getPlugIn(ArchivePlugIn.class).loadArchives();
+			final List<NexusArchiveWrapper> wrappers = App.getInstance().getPlugIn(ArchivePlugIn.class).getArchives();
 
 			gltfExporter.setGlTFExportMonitor(new GlTFExportMonitor() {
 				@Override
 				public void requestTextures(String textureId, ResourceBundle resourceBundle) {
-					final SearchResult r = find(vaults, textureId);
-					if (r == null) {
+					final IdxFileLink textureLink = find(wrappers, textureId);
+					if (textureLink == null) {
 						return;
 					}
 
-					final VaultReader archive = r.archive;
-					final IdxFileLink textureFile = r.fileLink;
-
 					try {
 						final TextureReader textureReader = TextureReader.buildDefault();
-						final TextureObject textureObject = textureReader.read(archive.getData(textureFile));
+						final TextureObject textureObject = textureReader.read(textureLink.getData());
 						final TextureImage origin = textureObject.getImage(0);
 
 						final List<TextureImage> images = new ArrayList<>();
@@ -238,7 +220,7 @@ final class ModelExporter implements Exporter {
 
 						for (int i = 0; i < images.size(); i++) {
 							final TextureImage image = images.get(i);
-							final String newfileName = String.format("%s.%d.png", textureFile.getNameWithoutFileEnding(), i);
+							final String newfileName = String.format("%s.%d.png", textureLink.getNameWithoutFileExtension(), i);
 							final BufferedImage bufferedImage = image.convertToBufferedImage();
 
 							final Path texDst = dst.resolve(Paths.get(newfileName));
@@ -249,10 +231,8 @@ final class ModelExporter implements Exporter {
 
 							resourceBundle.addTextureResource(new PathTextureResource(texDst));
 						}
-					} catch (final IOException e) {
-						throw new IllegalStateException(e);
-					} catch (final IdxEntryNotFound e2) {
-						// ignore
+					} catch (final IOException e1) {
+						throw new IllegalStateException(e1);
 					}
 				}
 
@@ -289,10 +269,10 @@ final class ModelExporter implements Exporter {
 		}
 
 		@Override
-		public void export(Model model, Path dstFolder, IdxFileLink fileLink) throws IOException {
+		public void export(Model model, Path dstFolder, IdxPath filePath) throws IOException {
 			final ModelDebugger debuger = ModelDebugger.createDefaultModelDebugger();
 			final Table table = debuger.debugModel(model);
-			final String modelName = fileLink.getName().substring(0, fileLink.getName().lastIndexOf('.'));
+			final String modelName = PathUtil.getNameWithoutExtension(filePath);
 			dstFolder = dstFolder.resolve(modelName + "_debug");
 			Files.createDirectories(dstFolder);
 			writeTable2CSV(modelName, table, dstFolder);