+++ src/main/java/nexusvault/cli/plugin/search/SearchPlugIn.java
@@ -7,13 +7,11 @@ import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
@@ -21,10 +19,10 @@ import java.util.stream.Stream;
 
 import nexusvault.archive.IdxDirectory;
 import nexusvault.archive.IdxEntry;
-import nexusvault.archive.IdxEntryNotAFile;
-import nexusvault.archive.IdxEntryNotFound;
+import nexusvault.archive.IdxEntryNotAFileException;
+import nexusvault.archive.IdxEntryNotFoundException;
 import nexusvault.archive.IdxFileLink;
-import nexusvault.archive.util.ArchivePath;
+import nexusvault.archive.IdxPath;
 import nexusvault.archive.util.IdxDirectoryTraverser;
 import nexusvault.archive.util.IdxEntryVisitor.EntryFilterResult;
 import nexusvault.archive.util.ReportingIdxFileCollector;
@@ -32,14 +30,11 @@ import nexusvault.cli.App;
 import nexusvault.cli.Command;
 import nexusvault.cli.plugin.AbstPlugIn;
 import nexusvault.cli.plugin.archive.ArchivePlugIn;
-import nexusvault.cli.plugin.archive.SourcedVaultReader;
+import nexusvault.cli.plugin.archive.NexusArchiveWrapper;
 
 public final class SearchPlugIn extends AbstPlugIn {
 
 	private final static String REPORT_FILE = "search_result.txt";
-
-	private final Map<Path, Set<IdxFileLink>> searchResults = new HashMap<>();
-	private boolean loadFromFile = true;
 
 	public SearchPlugIn() {
 		final List<Command> cmds = new ArrayList<>();
@@ -47,60 +42,63 @@ public final class SearchPlugIn extends AbstPlugIn {
 		setCommands(cmds);
 	}
 
-	public Map<Path, Set<IdxFileLink>> getLastSearchResults() {
-		if (loadFromFile) {
-			sendMsg("No active search found. Load previous search from " + REPORT_FILE);
-			return rebuildReportReferences();
+	public List<IdxPath> getLastSearchResults() {
+		try {
+			return loadReport();
+		} catch (final NoSuchFileException e1) {
+			sendMsg("No previous executed search found at");
+			sendMsg(e1.getMessage());
+			return Collections.emptyList();
+		} catch (final IOException e0) {
+			sendMsg("An error occured while loading search results from report file:");
+			sendMsg(e0.getClass().toString());
+			sendMsg(e0.getMessage());
+			return Collections.emptyList();
 		}
-		return Collections.unmodifiableMap(searchResults);
 	}
 
 	public void searchArchive(SearchRequest request) {
 		final ArchivePlugIn archivePlugIn = App.getInstance().getPlugIn(ArchivePlugIn.class);
-		final List<SourcedVaultReader> vaults = archivePlugIn.loadArchives();
-		if (vaults.isEmpty()) {
+		final List<NexusArchiveWrapper> wrappers = archivePlugIn.getArchives();
+		if (wrappers.isEmpty()) {
 			sendMsg("No vaults are loaded. Use 'help' to learn how to load them");
 			return;
 		}
 
-		searchResults.clear();
+		final List<IdxPath> searchResults = new LinkedList<>();
 
-		final ArchivePath path = archivePlugIn.getPathWithinArchives();
+		final IdxPath path = archivePlugIn.getPathWithinArchives();
 
 		int maxResults = request.getMaxResults();
 		int resultsFound = 0;
-		for (final SourcedVaultReader vault : vaults) {
-			sendMsg("Scanning archive: " + vault.getSource());
-			final IdxDirectory rootDirectory = vault.getReader().getRootFolder();
+		for (final NexusArchiveWrapper wrapper : wrappers) {
+			sendMsg("Scanning archive: " + wrapper.getSource());
+			final IdxDirectory rootDirectory = wrapper.getArchive().getRootDirectory();
 			if (path.isResolvable(rootDirectory)) {
 				final IdxEntry resolvedEntry = path.resolve(rootDirectory);
 				final Set<IdxFileLink> results = search(resolvedEntry, request.getPattern(), maxResults);
 				maxResults -= results.size();
 				resultsFound += results.size();
-				searchResults.put(vault.getSource(), Collections.unmodifiableSet(results));
+
+				results.stream().map(IdxFileLink::getPath).forEach(searchResults::add);
 			}
 		}
 
-		loadFromFile = false;
 		sendMsg(String.format("Found a total amount of %d files which match at least one search criteria", resultsFound));
 
 		try {
-			writeReport();
+			writeReport(searchResults);
 		} catch (final IOException e) {
 			sendMsg("An error occured:");
 			sendMsg(e.getMessage());
 		}
 	}
 
-	public void writeReport() throws IOException {
+	public void writeReport(Collection<IdxPath> searchResults) throws IOException {
 		sendMsg(String.format("Saving search results to file %s ...", REPORT_FILE));
 
 		final List<String> linesToWrite = new LinkedList<>();
-		for (final Entry<Path, Set<IdxFileLink>> searchEntry : searchResults.entrySet()) {
-			// linesToWrite.add(searchEntry.getKey().toString());
-			// linesToWrite.add(searchEntry.getValue().size() + "");
-			searchEntry.getValue().stream().map(f -> f.fullName()).sorted().forEach(linesToWrite::add);
-		}
+		searchResults.stream().map(IdxPath::getFullName).sorted().forEach(linesToWrite::add);
 
 		final Path reportFile = App.getInstance().getAppConfig().getReportFolder().resolve(REPORT_FILE);
 		Files.createDirectories(reportFile.getParent());
@@ -110,59 +108,19 @@ public final class SearchPlugIn extends AbstPlugIn {
 		sendMsg("... done.");
 	}
 
-	private Map<Path, Set<IdxFileLink>> rebuildReportReferences() {
-		final ArchivePlugIn archivePlugIn = App.getInstance().getPlugIn(ArchivePlugIn.class);
-		final List<SourcedVaultReader> vaults = archivePlugIn.loadArchives();
-		if (vaults.isEmpty()) {
-			sendMsg("No vaults are loaded. Use 'help' to learn how to load them");
-			return Collections.emptyMap();
-		}
-
-		final Map<Path, Set<IdxFileLink>> result = new HashMap<>();
-		for (final SourcedVaultReader vault : vaults) {
-			result.put(vault.getSource(), new HashSet<>());
-		}
-
-		List<String> reportSearchResults = null;
-		try {
-			reportSearchResults = loadReport();
-		} catch (final NoSuchFileException e1) {
-			sendMsg("No previous executed search found at");
-			sendMsg(e1.getMessage());
-			return Collections.emptyMap();
-		} catch (final IOException e0) {
-			sendMsg("An error occured while loading search results from report file:");
-			sendMsg(e0.getClass().toString());
-			sendMsg(e0.getMessage());
-			return Collections.emptyMap();
-		}
-
-		for (final String fileName : reportSearchResults) {
-			for (final SourcedVaultReader vault : vaults) {
-				final IdxFileLink file = findFile(vault.getReader().getRootFolder(), fileName);
-				if (file != null) {
-					result.get(vault.getSource()).add(file);
-					break;
-				}
-			}
-		}
-
-		return result;
-	}
-
-	private List<String> loadReport() throws IOException {
-		final List<String> result = new LinkedList<>();
+	private List<IdxPath> loadReport() throws IOException {
+		final List<IdxPath> result = new ArrayList<>();
 		final Path reportFile = App.getInstance().getAppConfig().getReportFolder().resolve(REPORT_FILE);
 		try (Stream<String> stream = Files.lines(reportFile, Charset.forName("UTF8"))) {
-			stream.forEach(result::add);
+			stream.map(IdxPath::createPathFrom).forEach(result::add);
 		}
 		return result;
 	}
 
-	private IdxFileLink findFile(IdxDirectory root, String fileName) {
+	private IdxFileLink findFile(IdxDirectory root, IdxPath path) {
 		try {
-			return root.getFile(fileName);
-		} catch (IdxEntryNotFound | IdxEntryNotAFile e) {
+			return path.resolve(root).asFile();
+		} catch (IdxEntryNotFoundException | IdxEntryNotAFileException e) {
 			return null;
 		}
 	}
@@ -170,7 +128,7 @@ public final class SearchPlugIn extends AbstPlugIn {
 	private Set<IdxFileLink> search(IdxEntry root, List<Pattern> regex, int maxResults) {
 		if (root.isFile()) {
 			for (final Pattern pattern : regex) {
-				if (pattern.matcher(root.fullName()).find()) {
+				if (pattern.matcher(root.getFullName()).find()) {
 					return Collections.singleton(root.asFile());
 				}
 			}
@@ -182,7 +140,7 @@ public final class SearchPlugIn extends AbstPlugIn {
 
 		final Predicate<IdxFileLink> fileTester = (file) -> {
 			for (final Pattern p : regex) {
-				if (p.matcher(file.fullName()).find()) {
+				if (p.matcher(file.getFullName()).find()) {
 					return true;
 				}
 			}